FIXES #30 #31 完成对maxdoc的分布式查询

+++ monad-api/src/main/scala/monad/api/internal/SearcherFacadeImpl.scala
@@ -20,7 +20,7 @@ class SearcherFacadeImpl(extractor: SearchResultExtractor, searcherQueue: Search
   private final val ONE_MINUTE = 60
   private val semaphore = new Semaphore(apiConfig.api.concurrentQuery)
 
-  def getDocumentNum: Int = searcherQueue.getDocumentNum
+  def getDocumentNum: Long = searcherQueue.getDocumentNum
 
   def facetSearch(searchRequest: SearchRequest): JsonObject = {
     extractor.extract(searchRequest, { request =>
