+++ monad-api/src/main/scala/monad/api/services/SearcherQueue.scala
@@ -24,7 +24,7 @@ trait SearcherQueue extends ServiceLifecycle {
    * @return 文档数目
    * @since 0.1
    */
-  def getDocumentNum: Int
+  def getDocumentNum: Long
 
   /**
    * 进行搜索
